remove dependency on DiseaseObserver
