+++ src/vivarium_ciff_sam/components/__init__.py
@@ -1,8 +1,7 @@
-from vivarium_ciff_sam.components.sqlns import SQLNSTreatment
 from vivarium_ciff_sam.components.intervention import SQLNSIntervention, WastingTreatmentIntervention
 from vivarium_ciff_sam.components.observers import (CategoricalRiskObserver, DisabilityObserver, DiseaseObserver,
                                                     MortalityObserver)
 from vivarium_ciff_sam.components.lbwsg import LowBirthWeight, ShortGestation
+from vivarium_ciff_sam.components.treatment import SQLNSTreatment, WastingTreatment
 from vivarium_ciff_sam.components.wasting import ChildWasting
-from vivarium_ciff_sam.components.wasting_treatment import WastingTreatment
 from vivarium_ciff_sam.components.x_factor import XFactorExposure
